Move the display bits to a separate library.
This caused issues when native code was attempted to be loaded in the frontend (zmq), even
though it wasn't using it.
